feat: use spline renderer to plot curved line graphs, full view impor… (#3919)
* feat: use spline renderer to plot curved line graphs, full view imporvements

* feat: increase min height for panel

* chore: move code to utils and plugins in uplot folder

* chore: update tooltip styles
