+++ frontend/src/container/GridCardLayout/GridCard/FullView/WidgetFullView.styles.scss
@@ -1,5 +1,5 @@
 .full-view-container {
-	height: 100%;
+	height: 80vh;
 	overflow-x: auto;
 	overflow-y: hidden;
 
@@ -21,13 +21,15 @@
 	.graph-manager-container {
 		height: calc(40% - 40px);
 
-		margin-top: 1.25rem;
-		display: flex;
-		align-items: flex-end;
-		overflow-x: scroll;
+		.graph-manager-header {
+			display: flex;
+			margin-bottom: 16px;
+		}
 
-		.filter-table-container {
-			flex-basis: 80%;
+		.legends-list-container {
+			width: 100%;
+			overflow: hidden;
+			overflow-y: auto;
 		}
 
 		.save-cancel-container {
