+++ frontend/src/container/GridCardLayout/config.ts
@@ -15,6 +15,6 @@ export const EMPTY_WIDGET_LAYOUT = {
 	i: PANEL_TYPES.EMPTY_WIDGET,
 	w: 6,
 	x: 0,
-	h: 2,
+	h: 3,
 	y: 0,
 };
