+++ frontend/src/container/NewDashboard/ComponentsSlider/index.tsx
@@ -45,7 +45,7 @@ function DashboardGraphSlider(): JSX.Element {
 							i: id,
 							w: 6,
 							x: 0,
-							h: 2,
+							h: 3,
 							y: 0,
 						},
 						...(layouts.filter((layout) => layout.i !== PANEL_TYPES.EMPTY_WIDGET) ||
