+++ frontend/src/lib/uPlotLib/uPlotLib.styles.scss
@@ -1,11 +1,29 @@
 .pointSquare {
-	width: 10px;
-	height: 10px;
-	background-color: white;
+	width: 12px;
+	height: 12px;
+	background-color: transparent;
 	border: 2px solid white;
 	box-sizing: border-box;
+	border-radius: 50%;
+}
+
+.tooltip-content-header {
+	margin-bottom: 8px;
+	font-size: 13px;
 }
 
 .tooltip-data-point {
 	font-size: 11px;
+}
+
+.tooltip-content {
+	display: flex;
+	align-items: center;
+	gap: 8px;
+	margin-bottom: 4px;
+
+	.pointSquare,
+	.tooltip-data-point {
+		font-size: 13px !important;
+	}
 }
