+++ frontend/src/lib/uPlotLib/utils/getGridColor.ts
@@ -1,0 +1,8 @@
+const getGridColor = (isDarkMode: boolean): string => {
+	if (isDarkMode) {
+		return 'rgba(231,233,237,0.2)';
+	}
+	return 'rgba(231,233,237,0.8)';
+};
+
+export default getGridColor;
